Refactor + move all helper funcs
