+++ corebehrt/main_causal/helper_scripts/get_stats.py
@@ -15,26 +15,13 @@ Outputs:
 """
 
 import logging
-from os.path import join
 
-import pandas as pd
-import torch
-
-from corebehrt.constants.causal.data import EXPOSURE_COL, PROBAS, PS_COL, TARGETS
-from corebehrt.constants.causal.paths import (
-    CALIBRATED_PREDICTIONS_FILE,
-    CRITERIA_FLAGS_FILE,
-    STATS_FILE_BINARY,
-    STATS_RAW_FILE_BINARY,
-    STATS_FILE_NUMERIC,
-    STATS_RAW_FILE_NUMERIC,
-)
-from corebehrt.constants.data import PID_COL
-from corebehrt.constants.paths import PID_FILE
 from corebehrt.functional.setup.args import get_args
 from corebehrt.main_causal.helper_scripts.helper.get_stat import (
-    StatConfig,
     analyze_cohort,
+    load_data,
+    print_stats,
+    save_stats,
 )
 from corebehrt.modules.setup.config import load_config
 from corebehrt.modules.setup.directory_causal import CausalDirectoryPreparer
@@ -69,72 +56,9 @@ def main(config_path: str):
         logger,
     )
 
-    print("================================================")
-    print("Criteria:")
-    print(criteria.head())
-    print("================================================")
     stats = analyze_cohort(criteria)
-    print("================================================")
-    print("Raw stats:")
-    print(stats["raw"]["binary"].head(20))
-    print(stats["raw"]["numeric"].head(20))
-    print("================================================")
-    print("Formatted stats:")
-    print(stats["formatted"]["binary"].head(20))
-    print(stats["formatted"]["numeric"].head(20))
-    stats["formatted"]["binary"].to_csv(join(save_path, STATS_FILE_BINARY), index=False)
-    stats["raw"]["binary"].to_csv(join(save_path, STATS_RAW_FILE_BINARY), index=False)
-    stats["formatted"]["numeric"].to_csv(
-        join(save_path, STATS_FILE_NUMERIC), index=False
-    )
-    stats["raw"]["numeric"].to_csv(join(save_path, STATS_RAW_FILE_NUMERIC), index=False)
-
-
-def load_data(
-    criteria_path: str,
-    cohort_path: str,
-    ps_calibrated_predictions_path: str,
-    outcome_model_path: str,
-    logger: logging.Logger,
-) -> pd.DataFrame:
-    """Load and merge cohort criteria, patient IDs, propensity scores, and predictions as needed."""
-
-    # Load main criteria DataFrame
-    logger.info("Loading criteria DataFrame")
-    criteria = pd.read_csv(join(criteria_path, CRITERIA_FLAGS_FILE))
-    logger.info(f"Loaded {len(criteria)} criteria")
-
-    # Optionally filter by patient IDs
-    if cohort_path:
-        pids = torch.load(join(cohort_path, PID_FILE))
-        logger.info(f"Loaded {len(pids)} patient IDs")
-        criteria = criteria[criteria[PID_COL].isin(pids)]
-        logger.info(f"Filtered criteria to {len(criteria)} patients")
-
-    # Optionally merge propensity scores and exposures
-    if ps_calibrated_predictions_path:
-        ps_path = join(ps_calibrated_predictions_path, CALIBRATED_PREDICTIONS_FILE)
-        ps_df = pd.read_csv(ps_path).rename(
-            columns={TARGETS: EXPOSURE_COL, PROBAS: PS_COL}
-        )
-        ps_df = _convert_to_int(ps_df, EXPOSURE_COL)
-        criteria = pd.merge(criteria, ps_df, on=PID_COL, how="left")
-        logger.info("Merged with propensity scores and exposures")
-
-    # Optionally merge predictions and targets
-    if outcome_model_path:
-        outcome_path = join(outcome_model_path, CALIBRATED_PREDICTIONS_FILE)
-        outcome_df = pd.read_csv(outcome_path)[[PID_COL, TARGETS]]
-        outcome_df = _convert_to_int(outcome_df, TARGETS)
-        criteria = pd.merge(criteria, outcome_df, on=PID_COL, how="left")
-        logger.info("Merged with predictions and targets")
-
-    return criteria
-
-
-def _convert_to_int(df: pd.DataFrame, col: str) -> pd.DataFrame:
-    df[col] = df[col].astype(int)
-    return df
+    print_stats(stats)
+    save_stats(stats, save_path)
 
 
 if __name__ == "__main__":
